requirements.txt
polymarket-apis>=0.3.0
supabase>=2.0.0,<2.10.0
python-dotenv>=1.0.0
httpx>=0.27.0
websockets>=12.0
pandas>=2.0.0
schedule>=1.2.0
loguru>=0.7.0
